refactor(brain): extract 4 modules from orchestrator_loop.py God Object
Issue #941: orchestrator_loop.py was 2461 lines with 15+ responsibilities.
Extracted into focused modules (~626 lines removed):

New modules:
- tool_result_summarizer.py: _summarize_tool_result, _prepare_tool_results_for_finalizer,
  _build_tool_success_summary, _count_items, _extract_count, _extract_field
- tool_plan_sanitizer.py: TOOL_REMAP, force_tool_plan, sanitize_tool_plan
- post_route_corrections.py: post_route_correction_email_send and email
  parsing helpers (looks_like_email_send_intent, extract_first_email, etc.)
- tool_param_builder.py: build_tool_params with gmail field aliasing

orchestrator_loop.py retains thin wrappers and backward-compat re-exports
so all existing callers continue to work without changes.

Closes #941
